pyproject.toml
[project]
name = "maya-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "maya-stubs>=0.4.2",
]
